+++ src/main/java/lotto/Application.java
@@ -1,9 +1,6 @@
 package lotto;
 
-import lotto.domain.LottoMachine;
-
-public class WebUILottoApplication {
-
+public class Application {
     public static void main(String[] args) {
         LottoMachine lottoMachine = new LottoMachine();
         lottoMachine.purchase();
